relocate helper tests to existing test project
