+++ AppwriteConsoleUnitTests/AppwriteConsoleUnitTests.csproj
@@ -21,4 +21,8 @@
     </PackageReference>
   </ItemGroup>
 
+    <ItemGroup>
+    <ProjectReference Include="..\AppwriteConsole.App\AppwriteConsole.App.csproj" />
+  </ItemGroup>
+
 </Project>
