+++ AppwriteConsoleUnitTests/CommandLineArgumentParserTests.cs
